+++ packages/Icon/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Ellipsis vertical icon
+- Character
+- Number,
+- Time and Date
+  [@nahumzs](https://github.com/nahumzs).
+
 ## [0.3.0] - 2020-04-29
 
 ### Removed